requirements.txt
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-admin-rangefilter==0.6.3
django-cleanup==5.1.0
django-import-export==2.4.0
django-session-security==2.6.6
django-storages==1.10.1
et-xmlfile==1.0.1
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
python-decouple==3.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
